# user_service/api/api.yml
openapi: 3.0.0
info:
  title: User Service API
  version: 1.0.0
paths:
  /register:
    post:
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationData'
      responses:
        '200':
          description: Пользователь успешно зарегистрирован
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '409':
          description: Пользователь с таким логином или email уже существует
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      error:
                        type: string
                        description: "Логин уже существует"
                  - type: object
                    properties:
                      error:
                        type: string
                        description: "Email уже существует"
        '500':
          description: Ошибка создания пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

  /login:
    post:
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthData'
      responses:
        '200':
          description: Успешная аутентификация
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '401':
          description: Неверный логин или пароль
        '500':
          description: Ошибка аутентификации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

  /profile:
    get:
      summary: Получение данных профиля пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthData'
      responses:
        '200':
          description: Данные профиля пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonData'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '401':
          description: Неверный логин или пароль
        '500':
          description: Ошибка получения профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

    put:
      summary: Обновление данных профиля пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileData'
      responses:
        '200':
          description: Профиль успешно обновлен
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '401':
          description: Неверный логин или пароль
        '500':
          description: Ошибка обновления профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

components:
  schemas:
    InternalServerError:
      type: object
      properties:
        error:
          type: string
          description: Ошибка
    
    BadRequestError:
      type: object
      properties:
        properties:
        error:
          type: string
          description: Ошибка

    RegistrationData:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
          format: password
        email:
          type: string
          format: email

    AuthData:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
          format: password

    UpdateProfileData:
      allOf:
        - $ref: '#/components/schemas/AuthData'
        - $ref: '#/components/schemas/PersonData'

    PersonData:
      type: object
      properties:
        name:
          type: string
          description: Имя
        surname:
          type: string
          description: Фамилия
        birth_date:
          type: string
          format: date
          description: Дата рождения
        email:
          type: string
          format: email
          description: Email
        phone_number:
          type: string
          description: Номер телефона
            
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT